+++ examples_new/sowfa_comparisons/yaw_comparisons/select_5_turb.py
@@ -22,14 +22,16 @@ import copy
 import pickle
 
 # Parameters
-num_turbines = 3
+num_turbines = 5
 sowfa_U0 = 8.0
-sowfa_TI = 0.1
-
-# Can chose between two x layouts
-# layout_x = (1000.0, 1756.0, 2512.0) 
-layout_x = (1000.0, 1882.0, 2764.0)
-layout_y = (1000.0, 1000.0, 1000.0)
+sowfa_TI = 0.1 # High = 0.1, low = 0.06
+layout_x = (1000.0, 1756.0, 2512.0, 3268.0, 4024.0)
+layout_y = (1000.0, 1000.0, 1000.0, 1000.0, 1000.0)
+yaw_cases_to_select = [
+    np.array([0.,0.,0.,0.,0.]),
+    np.array([25.,0.,0.,0.,0.]),
+    np.array([25.,25.,0.,0.,0.])
+]
 
 ## Grab certain hi-TI five simulations from saved SOWFA data set
 df_sowfa = pd.read_pickle('../sowfa_data_set/sowfa_data_set.p')
@@ -47,9 +49,14 @@ df_sowfa = df_sowfa[df_sowfa.sowfa_TI == sowfa_TI]
 df_sowfa = df_sowfa[df_sowfa.layout_x == layout_x]
 df_sowfa = df_sowfa[df_sowfa.layout_y == layout_y]
 
-# Sort by total sowfa power
-df_sowfa['total_sowfa_power'] = df_sowfa.power.apply(np.sum)
-df_sowfa = df_sowfa.sort_values('total_sowfa_power')
+# Limit to certain yaw cases
+def match_yaw(x):
+    for y in yaw_cases_to_select:
+        if np.array_equal(x,y):
+            return True
+    return False
+
+df_sowfa = df_sowfa[df_sowfa.yaw.apply(match_yaw)]
 
 # Load the saved FLORIS interfaces
 fi_dict = pickle.load( open( "../floris_models.p", "rb" ) )
@@ -61,7 +68,6 @@ for floris_label in fi_dict:
     df_sowfa[floris_label] = 0
     df_sowfa[floris_label] = df_sowfa[floris_label].astype(object)
     for i, row in df_sowfa.iterrows():
-
 
         # Match the layout, wind_speed and TI
         fi.reinitialize_flow_field(
@@ -117,6 +123,7 @@ fig, ax = plt.subplots(figsize=(7,4))
 case_names = df_sowfa.yaw.apply(lambda x: '/'.join(x.astype(int).astype(str)))
 sowfa_total = df_sowfa.power.apply(np.sum)
 ax.plot(sowfa_total,case_names,'ks-',label='SOWFA')
+
 # Plot the FLORIS results
 for floris_label in fi_dict:
     (fi, floris_color, floris_marker) = fi_dict[floris_label]
